+++ addon/services/syncer.js
@@ -11,6 +11,7 @@ import { SimplePredicate } from '../query/predicate';
 import { reloadLocalRecords, createLocalRecord } from '../utils/reload-local-records';
 import isModelInstance from '../utils/is-model-instance';
 import Queue from '../utils/queue';
+import { camelize, capitalize } from '../utils/string-functions';
 
 /**
   @class Syncer
@@ -202,6 +203,7 @@ export default Service.extend({
     @param {Ember.Array} [jobs] Array instances of `auditEntity` model for sync up.
     @param {Object} [options] Object with options for sync up.
     @param {Boolean} [options.continueOnError] If `true` continue sync up if an error occurred.
+    @param {Boolean} [options.useBatchUpdate] Do sync up through batch update or not.
     @return {Promise}
   */
   syncUp(jobs, options) {
@@ -222,7 +224,9 @@ export default Service.extend({
 
       return this.get('offlineStore').query(modelName, builder.build()).then((jobs) => {
         dexieService.set('queueSyncUpTotalWorksCount', jobs.get('length'));
-        return this._runJobs(store, jobs, options && options.continueOnError);
+        return options && options.useBatchUpdate ?
+          this._runJobsThroughBatch(store, jobs) :
+          this._runJobs(store, jobs.toArray(), options && options.continueOnError);
       });
     }
   },
@@ -319,6 +323,25 @@ export default Service.extend({
   },
 
   /**
+   * Return name of projection to sync up by model name.
+   * @method getSyncUpProjectionName
+   * @param {String} modelName Name of model to get sync up projection name.
+   * @returns Name of projection to sync up.
+   */
+  getSyncUpProjectionName(store, modelName) {
+    const modelClass = store.modelFor(modelName);
+
+    let projectionName = modelName.indexOf('-') > -1 ? modelName.substring(modelName.indexOf('-') + 1) : modelName;
+    projectionName = capitalize(camelize(projectionName)) + 'E';
+
+    if (modelClass.projections && modelClass.projections.get(projectionName)) {
+      return projectionName;
+    }
+
+    return null;
+  },
+
+  /**
   */
   _runJobs(store, jobs, continueOnError, jobCount) {
     let _this = this;
@@ -343,6 +366,47 @@ export default Service.extend({
         dexieService.set('queueSyncUpWorksCount', 0);
         resolve(executedJob);
       }
+    });
+  },
+
+  /**
+   * Starts syncing up through batch.
+   * @param {DS.Store} store
+   * @param {Array} jobs
+   */
+  _runJobsThroughBatch(store, jobs) {
+    let recordsToSyncUp = [];
+    let dexieService = getOwner(this).lookup('service:dexie');
+    dexieService.set('queueSyncUpWorksCount', jobs.get('length'));
+
+    return new RSVP.Promise((resolve, reject) => {
+      jobs.forEach(job => {
+        switch (job.get('operationType')) {
+          case 'INSERT':
+            recordsToSyncUp.push(this._getRecordToCreate(store, job));
+            break;
+          case 'UPDATE':
+            recordsToSyncUp.push(this._getRecordToUpdate(store, job));
+            break;
+          case 'DELETE':
+            recordsToSyncUp.push(this._getRecordToRemove(store, job));
+            break;
+          default:
+            throw new Error('Unsupported operation type.');
+        }
+      });
+
+      return RSVP.Promise.all(recordsToSyncUp).then(records => {
+        return store.batchUpdate(records).then(() => {
+          jobs.forEach(job => {
+            RSVP.all(job.get('auditFields').map(field => field.destroyRecord())).then(() => {
+              resolve(job.destroyRecord());
+            }, reject);
+          });
+        }).catch(error => {
+          reject(error);
+        });
+      });
     });
   },
 
@@ -377,19 +441,27 @@ export default Service.extend({
   },
 
   /**
-  */
-  _runCreatingJob(store, job) {
-    let record = store.peekRecord(job.get('objectType.name'), job.get('objectPrimaryKey'));
-    if (record) {
-      store.unloadRecord(record);
-    }
+   * Returns record to create for syncing up.
+   * @param {DS.Store} store Store.
+   * @param {DS.Model} job Job for sync up.
+   */
+  _getRecordToCreate(store, job) {
+    let record = store.peekRecord(job.get('objectType.name'), job.get('objectPrimaryKey')) ||
+      store.createRecord(job.get('objectType.name'), { id: job.get('objectPrimaryKey') });
 
-    record = store.createRecord(job.get('objectType.name'), { id: job.get('objectPrimaryKey') });
     record.set('isSyncingUp', true);
     record.set('isCreatedDuringSyncUp', true);
 
     return this._changesForRecord(store, job).then((changes) => {
       record.setProperties(changes);
+      return record;
+    });
+  },
+
+  /**
+  */
+  _runCreatingJob(store, job) {
+    return this._getRecordToCreate(store, job).then(record => {
       return record.save().then(() => {
         job.set('executionResult', 'Выполнено');
         return job.save();
@@ -402,23 +474,18 @@ export default Service.extend({
   },
 
   /**
-  */
-  _runUpdatingJob(store, job) {
+   * Returns record to update for syncing up.
+   * @param {DS.Store} store Store.
+   * @param {DS.Model} job Job for sync up.
+   */
+  _getRecordToUpdate(store, job) {
     let query = this._createQuery(store, job);
     return store.queryRecord(query.modelName, query).then((record) => {
       if (record) {
         record.set('isSyncingUp', true);
         return this._changesForRecord(store, job).then((changes) => {
           record.setProperties(changes);
-          return record.save().then(() => {
-            record.set('isUpdatedDuringSyncUp', true);
-            job.set('executionResult', 'Выполнено');
-            return job.save();
-          }).catch(reason => this.resolveServerError(job, reason)).finally(() => {
-            if (record) {
-              record.set('isSyncingUp', false);
-            }
-          });
+          return record;
         });
       } else {
         return this.resolveNotFoundRecord(job);
@@ -428,23 +495,51 @@ export default Service.extend({
 
   /**
   */
-  _runRemovingJob(store, job) {
+  _runUpdatingJob(store, job) {
+    return this._getRecordToUpdate(store, job).then(record => {
+      return record.save().then(() => {
+        record.set('isUpdatedDuringSyncUp', true);
+        job.set('executionResult', 'Выполнено');
+        return job.save();
+      }).catch(reason => this.resolveServerError(job, reason)).finally(() => {
+        if (record) {
+          record.set('isSyncingUp', false);
+        }
+      });
+    });
+  },
+
+  /**
+   * Returns record to remove for syncing up.
+   * @param {DS.Store} store Store.
+   * @param {DS.Model} job Job for sync up.
+   */
+  _getRecordToRemove(store, job) {
     let query = this._createQuery(store, job);
-    return store.queryRecord(query.modelName, query).then((record) => {
+    return store.queryRecord(query.modelName, query).then(record => {
       if (record) {
         record.set('isSyncingUp', true);
-        return record.destroyRecord().then(() => {
-          record.set('isDestroyedDuringSyncUp', true);
-          job.set('executionResult', 'Выполнено');
-          return job.save();
-        }).catch(reason => this.resolveServerError(job, reason)).finally(() => {
-          if (record) {
-            record.set('isSyncingUp', false);
-          }
-        });
+        record.deleteRecord();
+        return record;
       } else {
         return this.resolveNotFoundRecord(job);
       }
+    });
+  },
+
+  /**
+  */
+  _runRemovingJob(store, job) {
+    return this._getRecordToRemove(store, job).then(record => {
+      return record.destroyRecord().then(() => {
+        record.set('isDestroyedDuringSyncUp', true);
+        job.set('executionResult', 'Выполнено');
+        return job.save();
+      }).catch(reason => this.resolveServerError(job, reason)).finally(() => {
+        if (record) {
+          record.set('isSyncingUp', false);
+        }
+      });
     });
   },
 
@@ -578,14 +673,16 @@ export default Service.extend({
       let promises = [];
       let attributes = get(store.modelFor(job.get('objectType.name')), 'attributes');
       job.get('auditFields').forEach((auditField) => {
-        let descriptorField = auditField.get('field').split('@');
-        let field = descriptorField.shift();
-        let type = descriptorField.shift();
+        const [field, type] = auditField.get('field').split('@');
+
         if (type) {
-          let builder = new Builder(store, type).byId(auditField.get('newValue'));
-          promises.push(store.queryRecord(type, builder.build()).then((relationship) => {
-            changes[field] = relationship;
-          }));
+          let relationship = null;
+          if (auditField.get('newValue')) {
+            relationship = store.peekRecord(type, auditField.get('newValue')) ||
+              store.createExistingRecord(type, auditField.get('newValue'));
+          }
+
+          changes[field] = relationship;
         } else {
           let value = auditField.get('newValue');
           switch (attributes.get(field).type) {
@@ -625,7 +722,7 @@ export default Service.extend({
     } else {
       record.eachAttribute((name) => {
         let value = record.get(name);
-        if (value) {
+        if (value !== undefined) {
           changes[name] = [null, value];
         }
       });
@@ -634,7 +731,7 @@ export default Service.extend({
     let changedRelationships = record.changedBelongsTo();
     let snapshot = record._createSnapshot();
     record.eachRelationship((name, descriptor) => {
-      let changedRelationship = changedRelationships[name];
+      let changedRelationship = this.get('auditEnabled') ? changedRelationships[name] : [null, record.get(name)];
       if (changedRelationship && descriptor.kind === 'belongsTo') {
         let relationshipType = descriptor.type;
         if (descriptor.options && descriptor.options.polymorphic) {
@@ -655,18 +752,14 @@ export default Service.extend({
   /**
   */
   _createQuery(store, job) {
-    //TODO: Get projection for sync up from generated list of projections for models.
-    let modelName = job.get('objectType.name');
-    let modelClass = store.modelFor(modelName);
-    let projectionName = modelName.indexOf('-') > -1 ? modelName.substring(modelName.indexOf('-') + 1) : modelName;
-    projectionName = projectionName.camelize().capitalize() + 'E';
+    const modelName = job.get('objectType.name');
+    const projectionName = this.getSyncUpProjectionName(store, modelName);
 
-    let builder = new Builder(store).from(modelName);
-    if (modelClass.projections && modelClass.projections.get(projectionName)) {
+    let builder = new Builder(store).from(modelName).byId(job.get('objectPrimaryKey'));
+    if (projectionName) {
       builder = builder.selectByProjection(projectionName);
     }
 
-    builder = builder.byId(job.get('objectPrimaryKey'));
     let query = builder.build();
     query.useOnlineStore = true;
     return query;