+++ tests/dummy/app/services/store.js
@@ -3,6 +3,9 @@ import { computed } from '@ember/object';
 import DS from 'ember-data';
 import BaseStore from 'ember-flexberry-data/stores/base-store';
 import StoreMixin from 'ember-flexberry-data/mixins/store';
+import Ember from 'ember';
+import { Offline, Projection } from 'ember-flexberry-data';
+import OnlineStore from 'ember-flexberry-data/stores/online-store';
 
 export default BaseStore.extend({
   offlineModels: computed(function() {
@@ -53,8 +56,8 @@ export default BaseStore.extend({
         }
       },
     });
-    let owner = getOwner(this);
-    this.set('onlineStore', DS.Store.extend(StoreMixin).create(owner.ownerInjection()));
+    let owner = Ember.getOwner(this);
+    this.set('onlineStore', OnlineStore.extend(Projection.StoreMixin).create(owner.ownerInjection()));
     this._super(...arguments);
     this.set('offlineStore.dbName', 'TestDB');
   }