+++ tests/dummy/config/environment.js
@@ -54,8 +54,9 @@ module.exports = function(environment) {
 
     ENV.APP.rootElement = '#ember-testing';
 
-    // TODO: find proper way to pass arguments to tests
-    var testODataServiceURL = process.env.TEST_ODATA_SERVICE_URL;
+    // URL of the backend running in docker
+    var testODataServiceURL = 'http://localhost:6500/odata';
+
     ENV.APP.testODataService = !!testODataServiceURL;
     ENV.APP.testODataServiceURL = testODataServiceURL;
     ENV.APP.autoboot = false;